// apps/frontend/src/app/api/models/device-create-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DeviceCapability } from '../models/device-capability';
import { DeviceTransport } from '../models/device-transport';
import { DeviceType } from '../models/device-type';
export interface DeviceCreateDto {
  capabilities?: Array<DeviceCapability>;
  description?: string;
  deviceId: string;
  displayName: string;
  location?: string;
  role?: string;
  transport?: DeviceTransport;
  type: DeviceType;
}