// apps/frontend/src/app/api/models/device-update-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DeviceCapability } from '../models/device-capability';
import { DeviceStatus } from '../models/device-status';
import { DeviceTransport } from '../models/device-transport';
import { DeviceType } from '../models/device-type';
export interface DeviceUpdateDto {
  capabilities?: Array<DeviceCapability>;
  description?: string;
  displayName: string;
  enabled: boolean;
  location?: string;
  role?: string;
  status: DeviceStatus;
  transport: DeviceTransport;
  type: DeviceType;
  version: number;
}